Move analysis helper methods to DirectCandidateGenerator
